+++ lib/kitchen/directions/bake_lists_with_para.rb
@@ -3,13 +3,13 @@
 module Kitchen
   module Directions
     # Bakes lists with paragraphs
-    # Should be used for books created by Adaptarr to put list text directly to <li>
+    # Should be used for books created by Adaptarr to put list content directly to <li>
     #
     module BakeListsWithPara
       def self.v1(book:)
         book.search('li').each do |item|
-          item_content = item.first('p')&.text
-          item.replace_children with: item_content unless item_content.nil?
+          item_para = item.first('p')&.cut
+          item.replace_children with: item_para.children unless item_para.nil?
         end
       end
     end